One more

+++ src/kernels/1_naive.cuh
@@ -12,14 +12,14 @@ MxK * KxN = MxN
 
 */
 
-__global__ void sgemm_naive(int M, int N, int K, float alpha, const float *A,
-                            const float *B, float beta, float *C) {
+__global__ void sgemm_naive(int M, int N, int K, double alpha, const double *A,
+                            const double *B, double beta, double *C) {
   const uint x = blockIdx.x * blockDim.x + threadIdx.x;
   const uint y = blockIdx.y * blockDim.y + threadIdx.y;
 
   // if statement is necessary to make things work under tile quantization
   if (x < M && y < N) {
-    float tmp = 0.0;
+    double tmp = 0.0;
     for (int i = 0; i < K; ++i) {
       tmp += A[x * K + i] * B[i * N + y];
     }
